# ICICI_Server/doc/swagger.json
{
    "info": {
        "version": "v0.0.1",
        "title": "ICICI",
        "description": "To perform any operation, you will need to first generate an access token. \n Tokens are generated by invoking the `/auth` API. \n The response is as follows, `Bearer <access-token>`. \n Copy the response and append it to all subsequent APIs where `Authorization` is required."
    },
...
    "paths": {
        "/api/users/": {
            "get": {
                "summary": "This endpoint will get user details.",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "Successfully get data",
                        "schema": {
                            "$ref": "#/responses/getUser"
                        },
                        "headers": {
                            "Access-Control-Allow-Origin": {
                                "type": "string",
                                "description": "URI that may access the resource"
                            }
                        }
                    },
                    "400": {
                        "description": "Validation error"
                    },
                    "500": {
                        "description": "Internal server error."
                    }
                },
                "parameters": [
                    {
                        "required": false,
                        "description": "Please enter maker approval status. (i.e. true/false )",
                        "name": "makerApproval",
                        "in": "query"
                    },
                    {
                        "required": false,
                        "description": "Please enter checker approval status. (i.e. true/false )",
                        "name": "checkerApproval",
                        "in": "query"
                    },
                    {
                        "required": false,
                        "description": "Please enter username",
                        "name": "username",
                        "in": "query"
                    }
                ],
                "tags": [
                    "User"
                ]
            }
        },
        "/api/user/login": {
            "post": {
                "summary": "This endpoint will authenticate user.",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "Successfully created",
                        "schema": {
                            "$ref": "#/responses/login"
                        },
                        "headers": {
                            "Access-Control-Allow-Origin": {
                                "type": "string",
                                "description": "URI that may access the resource"
                            }
                        }
                    },
                    "400": {
                        "description": "Validation error"
                    },
                    "500": {
                        "description": "Internal server error."
                    }
                },
                "parameters": [
                    {
                        "schema": {
                            "$ref": "#/definitions/login"
                        },
                        "required": true,
                        "description": "No Description.",
                        "name": "body",
                        "in": "body"
                    }
                ],
                "tags": [
                    "User"
                ]
            }
        },
        "/api/user/logout": {
            "post": {
                "summary": "This endpoint will expire session.",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "Successfully logout",
                        "schema": {
                            "$ref": "#/responses/logOut"
                        },
                        "headers": {
                            "Access-Control-Allow-Origin": {
                                "type": "string",
                                "description": "URI that may access the resource"
                            }
                        }
                    },
                    "400": {
                        "description": "Validation error"
                    },
                    "500": {
                        "description": "Internal server error."
                    }
                },
                "parameters": [],
                "tags": [
                    "User"
                ]
            }
        },
        "/api/user/register": {
            "post": {
                "summary": "This endpoint will add new user.",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "Successfully created",
                        "schema": {
                            "$ref": "#/responses/userRegistration"
                        },
                        "headers": {
                            "Access-Control-Allow-Origin": {
                                "type": "string",
                                "description": "URI that may access the resource"
                            }
                        }
                    },
                    "400": {
                        "description": "Validation error"
                    },
                    "500": {
                        "description": "Internal server error."
                    }
                },
                "parameters": [
                    {
                        "schema": {
                            "$ref": "#/definitions/userRegistration"
                        },
                        "required": true,
                        "description": "No Description.",
                        "name": "body",
                        "in": "body"
                    }
                ],
                "tags": [
                    "User"
                ]
            }
        },
        "/api/user/": {
            "get": {
                "summary": "This endpoint will get user details.",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "Successfully get data",
                        "schema": {
                            "$ref": "#/responses/getUser"
                        },
                        "headers": {
                            "Access-Control-Allow-Origin": {
                                "type": "string",
                                "description": "URI that may access the resource"
                            }
                        }
                    },
                    "400": {
                        "description": "Validation error"
                    },
                    "500": {
                        "description": "Internal server error."
                    }
                },
                "parameters": [
                    {
                        "required": true,
                        "description": "Please enter username.",
                        "name": "username",
                        "in": "query"
                    }
                ],
                "tags": [
                    "User"
                ]
            },
            "put": {
                "summary": "This endpoint will update the existing user.",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "Successfully created",
                        "schema": {
                            "$ref": "#/responses/updateUser"
                        },
                        "headers": {
                            "Access-Control-Allow-Origin": {
                                "type": "string",
                                "description": "URI that may access the resource"
                            }
                        }
                    },
                    "400": {
                        "description": "Validation error"
                    },
                    "500": {
                        "description": "Internal server error."
                    }
                },
                "parameters": [
                    {
                        "schema": {
                            "$ref": "#/definitions/updateUser"
                        },
                        "required": true,
                        "description": "No Description.",
                        "name": "body",
                        "in": "body"
                    }
                ],
                "tags": [
                    "User"
                ]
            }
        },
...
        "/api/user/changePassword": {
            "post": {
                "summary": "This endpoint will change password.",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "Successfully updated.",
                        "schema": {
                            "$ref": "#/responses/changePassword"
                        },
                        "headers": {
                            "Access-Control-Allow-Origin": {
                                "type": "string",
                                "description": "URI that may access the resource"
                            }
                        }
                    },
                    "400": {
                        "description": "Validation error"
                    },
                    "500": {
                        "description": "Internal server error."
                    }
                },
                "parameters": [
                    {
                        "schema": {
                            "$ref": "#/definitions/changePassword"
                        },
                        "required": true,
                        "description": "",
                        "name": "body",
                        "in": "body"
                    }
                ],
                "tags": [
                    "User"
                ]
            }
        },
        "/api/user/approveByMaker": {
            "post": {
                "summary": "This endpoint will approve the user by maker role.",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "Successfully updated.",
                        "schema": {
                            "$ref": "#/responses/approveByMaker"
                        },
                        "headers": {
                            "Access-Control-Allow-Origin": {
                                "type": "string",
                                "description": "URI that may access the resource"
                            }
                        }
                    },
                    "400": {
                        "description": "Validation error"
                    },
                    "500": {
                        "description": "Internal server error."
                    }
                },
                "parameters": [
                    {
                        "schema": {
                            "$ref": "#/definitions/approveByMaker"
                        },
                        "required": true,
                        "description": "",
                        "name": "body",
                        "in": "body"
                    }
                ],
                "tags": [
                    "User"
                ]
            }
        },
        "/api/user/approveByChecker": {
            "post": {
                "summary": "This endpoint will approve the user by checker role.",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "Successfully updated.",
                        "schema": {
                            "$ref": "#/responses/approveByMaker"
                        },
                        "headers": {
                            "Access-Control-Allow-Origin": {
                                "type": "string",
                                "description": "URI that may access the resource"
                            }
                        }
                    },
                    "400": {
                        "description": "Validation error"
                    },
                    "500": {
                        "description": "Internal server error."
                    }
                },
                "parameters": [
                    {
                        "schema": {
                            "$ref": "#/definitions/approveByChecker"
                        },
                        "required": true,
                        "description": "",
                        "name": "body",
                        "in": "body"
                    }
                ],
                "tags": [
                    "User"
                ]
            }
        },
        "/api/user/forgetPassword": {
            "put": {
                "summary": "This endpoint will update password.",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "Successfully created",
                        "schema": {
                            "$ref": "#/responses/updatePassword"
                        },
                        "headers": {
                            "Access-Control-Allow-Origin": {
                                "type": "string",
                                "description": "URI that may access the resource"
                            }
                        }
                    },
                    "400": {
                        "description": "Validation error"
                    },
                    "500": {
                        "description": "Internal server error."
                    }
                },
                "parameters": [
                    {
                        "schema": {
                            "$ref": "#/definitions/updatePassword"
                        },
                        "required": true,
                        "description": "No Description.",
                        "name": "body",
                        "in": "body"
                    }
                ],
                "tags": [
                    "User"
                ]
            }
        },
        "/api/user/deleteUser": {
            "delete": {
                "summary": "This endpoint will delete user.",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "Successfully created",
                        "schema": {
                            "$ref": "#/responses/deleteUser"
                        },
                        "headers": {
                            "Access-Control-Allow-Origin": {
                                "type": "string",
                                "description": "URI that may access the resource"
                            }
                        }
                    },
                    "400": {
                        "description": "Validation error"
                    },
                    "500": {
                        "description": "Internal server error."
                    }
                },
                "parameters": [
                    {
                        "required": true,
                        "description": "No Description.",
                        "name": "username",
                        "in": "query"
                    }
                ],
                "tags": [
                    "User"
                ]
            }
        },
...
    },
...
    "securityDefinitions": {
        "access_token": {
            "description": "Authorization For Transaction APIs",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
    "definitions": {
        "login": {
            "required": [
                "username",
                "password"
            ],
            "properties": {
                "username": {
                    "type": "string",
                    "description": "Enter user name.",
                    "format": "string"
                },
                "password": {
                    "type": "string",
                    "description": "",
                    "format": "string"
                }
            }
        },
        "userRegistration": {
            "required": [],
            "properties": {
                "organisationName": {
                    "type": "string",
                    "description": "Enter user name.",
                    "format": "string"
                },
                "vanNo": {
                    "type": "string",
                    "description": "",
                    "format": "string"
                },
                "IFSCCode": {
                    "type": "string",
                    "description": "",
                    "format": "string"
                },
                "accountNumber": {
                    "type": "string",
                    "description": "",
                    "format": "string"
                },
                "creditAccountNumber": {
                    "type": "string",
                    "description": "",
                    "format": "string"
                },
                "businessSpocName": {
                    "type": "string",
                    "description": "",
                    "format": "string"
                },
                "businessSpocEmails": {
                    "type": "array",
                    "items": {
                        "type": "object",
                        "properties": {
                            "businessSpocEmail": {
                                "type": "string",
                                "description": ""
                            }
                        }
                    },
                    "description": "List of Terms and conditions for the disbursal.",
                    "format": "string"
                },
                "businessSpocMobileNumber": {
                    "type": "string",
                    "description": "",
                    "format": "string"
                },
                "ITSpocName": {
                    "type": "string",
                    "description": "",
                    "format": "string"
                },
                "ITSpocEmails": {
                    "type": "array",
                    "items": {
                        "type": "object",
                        "properties": {
                            "ITSpocEmail": {
                                "type": "string",
                                "description": ""
                            }
                        }
                    },
                    "description": "List of Terms and conditions for the disbursal.",
                    "format": "string"
                },
                "ITSpocMobileNumber": {
                    "type": "string",
                    "description": "",
                    "format": "string"
                },
                "webService": {
                    "type": "string",
                    "description": "",
                    "format": "string"
                },
                "uatPublicIP": {
                    "type": "string",
                    "description": "",
                    "format": "string"
                },
                "uatPortNo": {
                    "type": "string",
                    "description": "",
                    "format": "string"
                },
                "livePublicIP": {
                    "type": "string",
                    "description": "",
                    "format": "string"
                },
                "livePortNo": {
                    "type": "string",
                    "description": "",
                    "format": "string"
                }
            }
        },
        "changePassword": {
            "required": [
                "username",
                "oldPassword",
                "newPassword",
                "confirmPassword"
            ],
            "properties": {
                "username": {
                    "type": "string",
                    "description": "",
                    "format": "string"
                },
                "oldPassword": {
                    "type": "string",
                    "description": "",
                    "format": "string"
                },
                "newPassword": {
                    "type": "string",
                    "description": "",
                    "format": "string"
                },
                "confirmPassword": {
                    "type": "string",
                    "description": "",
                    "format": "string"
                }
            }
        },
...
        "approveByMaker": {
            "required": [],
            "properties": {
                "projectId": {
                    "type": "string",
                    "description": "Enter project ID. ",
                    "format": "string"
                },
                "makerApproval": {
                    "type": "string",
                    "description": "Enter approval status (true/false). ",
                    "format": "string"
                },
                "status": {
                    "type": "string",
                    "description": "Enter username.",
                    "format": "string"
                },
                "username": {
                    "type": "string",
                    "description": "Enter username.",
                    "format": "string"
                },
                "createdBy": {
                    "type": "string",
                    "description": "Enter createdBy (i.e. username).",
                    "format": "string"
                },
                "orgName": {
                    "type": "string",
                    "description": "Enter username.",
                    "format": "string"
                }
            }
        },
        "approveByChecker": {
            "required": [],
            "properties": {
                "projectId": {
                    "type": "string",
                    "description": "Enter project ID. ",
                    "format": "string"
                },
                "repoName": {
                    "type": "string",
                    "description": "Enter repo name (Ex. ProjectName). ",
                    "format": "string"
                },
                "checkerApproval": {
                    "type": "string",
                    "description": "Enter approval status (true/false). ",
                    "format": "string"
                },
                "status": {
                    "type": "string",
                    "description": "Enter username.",
                    "format": "string"
                },
                "username": {
                    "type": "string",
                    "description": "Enter username.",
                    "format": "string"
                },
                "createdBy": {
                    "type": "string",
                    "description": "Enter createdBy (i.e. username).",
                    "format": "string"
                },
                "orgName": {
                    "type": "string",
                    "description": "Enter username.",
                    "format": "string"
                }
            }
        },
        "updateUser": {
            "required": [
                "username"
            ],
            "properties": {
                "username": {
                    "type": "string",
                    "description": "Enter user name.",
                    "format": "string"
                },
                "firstName": {
                    "type": "string",
                    "description": "",
                    "format": "string"
                },
                "lastName": {
                    "type": "string",
                    "description": "",
                    "format": "string"
                },
                "mobileNumber": {
                    "type": "string",
                    "description": "",
                    "format": "string"
                },
                "email": {
                    "type": "string",
                    "description": "",
                    "format": "string"
                }
            }
        },
...
    },
    "responses": {
        "getUser": {
            "required": [],
            "properties": {
                "username": {
                    "type": "string",
                    "description": "Enter user name.",
                    "format": "string"
                },
                "firstName": {
                    "type": "string",
                    "description": "",
                    "format": "string"
                },
                "lastName": {
                    "type": "string",
                    "description": "",
                    "format": "string"
                },
                "mobileNumber": {
                    "type": "string",
                    "description": "",
                    "format": "string"
                },
                "email": {
                    "type": "string",
                    "description": "",
                    "format": "string"
                }
            }
        },
        "login": {
            "required": [],
            "properties": {
                "token": {
                    "type": "string",
                    "description": "You will get token.",
                    "format": "string"
                }
            }
        },
        "logOut": {
            "required": [],
            "properties": {
                "token": {
                    "type": "string",
                    "description": "",
                    "format": "string"
                }
            }
        },
        "userRegistration": {
            "required": [],
            "properties": {
                "message": {
                    "type": "string",
                    "description": "You will get message.",
                    "format": "string"
                }
            }
        },
        "changePassword": {
            "required": [],
            "properties": {
                "message": {
                    "type": "string",
                    "description": "You will get message.",
                    "format": "string"
                }
            }
        },
        "updatePassword": {
            "required": [],
            "properties": {
                "message": {
                    "type": "string",
                    "description": "You will get message.",
                    "format": "string"
                }
            }
        },
        "updateUser": {
            "required": [],
            "properties": {
                "message": {
                    "type": "string",
                    "description": "You will get message.",
                    "format": "string"
                }
            }
        },
...
    },
...
    "security": [
        {
            "access_token": []
        }
    ],
    "swagger": "2.0"
}